requirements.txt
django==1.11
Pillow==5
django_rest_framework
django-rest-swagger